+++ spec/features/draws/draw_group_report_spec.rb
@@ -3,6 +3,8 @@
 require 'rails_helper'
 
 RSpec.feature 'Draw group report' do
+  include GroupsHelper # for status-displaying
+
   let(:draw) { FactoryGirl.create(:draw, status: 'pre_lottery') }
   let!(:groups) { create_groups(draw: draw, statuses: %w(full open locked)) }
 
@@ -42,11 +44,12 @@ RSpec.feature 'Draw group report' do
 
   def page_has_group_report(page:, group:, edit: false)
     edit_assert_method = edit ? :assert_selector : :refute_selector
+    status_text = display_group_status(group)
     within(".group-report tr#group-#{group.id}") do
       page.assert_selector(:css, 'td[data-role="group-leader"]',
                            text: group.leader.full_name) &&
         page.assert_selector(:css, 'td[data-role="group-status"]',
-                             text: group.status.capitalize) &&
+                             text: status_text) &&
         page.send(edit_assert_method, :css, 'a.button', text: 'Edit')
     end
   end